Feature : 사용자 - 서버 클라이언트 연동을 위한 기본 파일 수정 (Issue #3)
