+++ src/main/java/com/eojin/shopeasy_back/user/dto/LoginRequestDto.java
@@ -1,0 +1,11 @@
+package com.eojin.shopeasy_back.user.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class LoginRequestDto {
+    private String username;
+    private String password;
+}
