+++ src/main/java/com/eojin/shopeasy_back/user/dto/LoginResponseDto.java
@@ -1,0 +1,14 @@
+package com.eojin.shopeasy_back.user.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class LoginResponseDto {
+    private String message;
+
+    public LoginResponseDto(String message) {
+        this.message = message;
+    }
+}
