+++ src/main/java/com/eojin/shopeasy_back/user/dto/UpdateUserRequestDto.java
@@ -1,26 +1,23 @@
-package com.eojin.shopeasy_back.user;
+package com.eojin.shopeasy_back.user.dto;
 
+import com.eojin.shopeasy_back.user.entity.User;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class CreateUserRequestDto {
-private String username;
+public class UpdateUserRequestDto {
 private String nickname;
-private String email;
 private String profile_url;
 private String password;
 
 public User toEntity() {
     return User.builder()
-                   .username(username)
                    .nickname(nickname)
-                   .email(email)
                    .profile_url(profile_url)
                    .password(password)
                    .build();
 }
 }
+
