+++ src/main/java/com/eojin/shopeasy_back/user/repository/UserRepository.java
@@ -1,8 +1,13 @@
-package com.eojin.shopeasy_back.user;
+package com.eojin.shopeasy_back.user.repository;
 
+import com.eojin.shopeasy_back.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByEmail(String email);
 }
